fastapi-app/requirements.txt
pymongo
fastapi
python-dotenv
uvicorn
bson